// apps/web-svelte/src/lib/api/generated/model/pluginSearchField.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PluginSearchFieldType } from './pluginSearchFieldType';

export interface PluginSearchField {
  key: string;
  label: string;
  type: PluginSearchFieldType;
  required: boolean;
  /** @nullable */
  placeholder?: string | null;
  /** @nullable */
  help?: string | null;
}